Split up Entity and Commentable.
